quite a lot of fixups, and a lot of docs

+++ src/daemon/daemon.ts
@@ -1,17 +1,18 @@
 import { ConsoleHandler } from '@debugr/console';
 import { Logger, LogLevel } from '@debugr/core';
+import { ZodError } from 'zod';
 import {
   Config,
   DaemonStatus,
-  DaemonUpgradeReply,
   getNodesockdVersion,
-  loadConfig,
+  loadConfig, WorkerRestartReply,
 } from '../common';
 import { IpcPeer, UnixSocketIpcServer, UnixSocketIpcTransport } from '../ipc';
-import { sleep } from '../utils';
+import { PromiseAborted, sleep } from '../utils';
 import { DevServer } from './devServer';
 import { ProcessManager } from './processManager';
 import { DaemonIpcIncomingMap } from './types';
+import { lte as isCurrentVersion } from 'semver';
 
 export class Daemon {
   private readonly configPath: string;
@@ -22,6 +23,7 @@ export class Daemon {
   private readonly devServer?: DevServer;
   private readonly version: string;
   private readonly startTs: number;
+  private terminating: boolean = false;
 
   constructor(configPath: string, config: Config, devServerPort?: number) {
     this.configPath = configPath;
@@ -42,14 +44,27 @@ export class Daemon {
     devServerPort && (this.devServer = new DevServer(devServerPort));
     this.handleSignal = this.handleSignal.bind(this);
     this.handleIpcConnection = this.handleIpcConnection.bind(this);
+    this.handleUncaughtError = this.handleUncaughtError.bind(this);
   }
 
   async run(): Promise<void> {
+    try {
+      await this.start();
+    } catch (e) {
+      await this.handleUncaughtError(e);
+    }
+  }
+
+  private async start(): Promise<void> {
     this.logger.warning('Daemon starting...');
 
     process.title = `${this.config.name} daemon`;
     process.on('SIGTERM', this.handleSignal);
     process.on('SIGINT', this.handleSignal);
+    process.on('SIGHUP', this.handleSignal);
+    process.on('uncaughtException', this.handleUncaughtError);
+    process.on('unhandledRejection', this.handleUncaughtError);
+    process.umask(0o007);
 
     this.ipc.on('connection', this.handleIpcConnection);
     await this.ipc.start();
@@ -77,21 +92,45 @@ export class Daemon {
     };
   }
 
-  async reloadConfig(): Promise<void> {
+  async reloadConfig(catchErrors: boolean = false): Promise<void> {
     this.logger.info('Reloading daemon config...');
-    const [config] = await loadConfig('/', this.configPath);
-    this.config = config;
-    process.title = `${config.name} daemon`;
-    await this.pm.setConfig(config);
-    this.logger.info('Daemon config reloaded successfully');
+
+    try {
+      const [config] = await loadConfig('/', this.configPath);
+      this.config = config;
+      process.title = `${config.name} daemon`;
+      await this.pm.setConfig(config);
+      this.logger.info('Daemon config reloaded successfully');
+    } catch (e) {
+      if (!catchErrors || !(e instanceof ZodError)) {
+        throw e;
+      }
+
+      const msg = e.errors.length > 1 ? `\n - ${e.errors.join('\n - ')}` : ` ${e.errors.join('')}`;
+      this.logger.error(`Failed to reload config file:${msg}`);
+    }
   }
 
-  async upgrade(version: string): Promise<DaemonUpgradeReply> {
-    if (version === this.version) {
-      return {
-        upgrading: false,
-        pid: process.pid,
-      };
+  async terminate(detach: boolean = false): Promise<void> {
+    if (this.terminating) {
+      return;
+    }
+
+    this.logger.warning('Daemon terminating...');
+    process.off('SIGTERM', this.handleSignal);
+    process.off('SIGINT', this.handleSignal);
+    process.off('SIGHUP', this.handleSignal);
+    this.terminating = true;
+
+    await this.pm.stop(detach);
+    await this.ipc.close();
+    await this.devServer?.close();
+  }
+
+  private async handleRestart(suspended?: boolean, version?: string): Promise<WorkerRestartReply> {
+    if (version === undefined || isCurrentVersion(version, this.version)) {
+      await this.pm.restart(suspended);
+      return { pid: process.pid };
     } else {
       this.logger.warning(`Upgrading daemon to version ${version}...`);
       sleep(500).then(() => this.terminate(true));
@@ -103,17 +142,11 @@ export class Daemon {
     }
   }
 
-  async terminate(detach: boolean = false): Promise<void> {
-    this.logger.warning('Daemon terminating...');
-    process.off('SIGTERM', this.handleSignal);
-    process.off('SIGINT', this.handleSignal);
+  private async handleSignal(signal: 'SIGTERM' | 'SIGINT' | 'SIGHUP'): Promise<void> {
+    if (signal === 'SIGHUP') {
+      return this.reloadConfig(true);
+    }
 
-    await this.pm.stop(detach);
-    await this.ipc.close();
-    await this.devServer?.close();
-  }
-
-  private async handleSignal(signal: 'SIGTERM' | 'SIGINT'): Promise<void> {
     await this.terminate();
     process.kill(process.pid, signal);
   }
@@ -122,11 +155,12 @@ export class Daemon {
     const peer = new IpcPeer<{}, DaemonIpcIncomingMap>(socket);
 
     peer.setRequestHandler('status', async () => this.getStatus());
-
     peer.setMessageHandler('online', async (data) => this.pm.handleOnline(peer, data));
     peer.setMessageHandler('broken', async (data) => this.pm.handleBroken(peer, data));
     peer.setRequestHandler('start-workers', async ({ suspended }) => this.pm.start(suspended));
-    peer.setRequestHandler('restart-workers', async ({ suspended }) => this.pm.restart(suspended));
+    peer.setRequestHandler('restart-workers', async ({ suspended, version }) => {
+      return this.handleRestart(suspended, version)
+    });
     peer.setRequestHandler('resume-workers', async () => this.pm.resume());
     peer.setRequestHandler('stop-workers', async () => this.pm.stop());
     peer.setRequestHandler('set-worker-count', async ({ count }) => this.pm.setWorkerCount(count));
@@ -139,12 +173,25 @@ export class Daemon {
       return this.pm.sendAppRequest(request, data, workers);
     });
 
-    peer.setRequestHandler('upgrade', async ({ version }) => this.upgrade(version));
+    peer.setRequestHandler('reload', async () => this.reloadConfig());
     peer.setRequestHandler('terminate', async () => {
       sleep(500).then(() => this.terminate());
       return { pid: process.pid };
     });
 
     await peer.run();
+  }
+
+  private async handleUncaughtError(err: unknown): Promise<void> {
+    if (this.terminating && err instanceof PromiseAborted) {
+      return;
+    } else if (err instanceof Error) {
+      this.logger.fatal('Uncaught application error', err);
+    } else {
+      this.logger.fatal('Uncaught application error', { error: err });
+    }
+
+    await this.terminate();
+    process.exit(1);
   }
 }
